Bumping database-proxy-1.4.3 + other dependency updates (#2)
* Ignore eclipse specific meta files

* Get the project compile

* Lock the version of the maven plugins

* Lock the java version

* Bump junit to 1.12

* bump jcl-over-slf4j to the newest 1.6 version: 1.6.6

* Remove unused version property (logback-ext-spring)

* Bump hibernate to the newest 4.3 version (4.3.11)

* Bump aspectj to the latest version of the 1.6 series (1.6.12)

* Bump datasource-proxy to the latest 1.3 version (1.3.3)

* bump jcl-over-slf4j to 1.7.25

* Bump logback to version 1.2.3

* Bump commons-lang to version 2.6

* Bump mockito to 1.10.19

* Bump datasource-proxy to version 1.4.1

* Bump datasource-proxy to 1.4.2

* Bump datasource-proxy to verison 1.4.3

+++ src/main/java/com/vladmihalcea/sql/SQLStatementCountValidator.java
@@ -48,9 +48,9 @@ public class SQLStatementCountValidator {
      *
      * @param expectedSelectCount expected select statement count
      */
-    public static void assertSelectCount(int expectedSelectCount) {
+    public static void assertSelectCount(long expectedSelectCount) {
         QueryCount queryCount = QueryCountHolder.getGrandTotal();
-        int recordedSelectCount = queryCount.getSelect();
+        long recordedSelectCount = queryCount.getSelect();
         if (expectedSelectCount != recordedSelectCount) {
             throw new SQLSelectCountMismatchException(expectedSelectCount, recordedSelectCount);
         }
@@ -61,9 +61,9 @@ public class SQLStatementCountValidator {
      *
      * @param expectedInsertCount expected insert statement count
      */
-    public static void assertInsertCount(int expectedInsertCount) {
+    public static void assertInsertCount(long expectedInsertCount) {
         QueryCount queryCount = QueryCountHolder.getGrandTotal();
-        int recordedInsertCount = queryCount.getInsert();
+        long recordedInsertCount = queryCount.getInsert();
         if (expectedInsertCount != recordedInsertCount) {
             throw new SQLInsertCountMismatchException(expectedInsertCount, recordedInsertCount);
         }
@@ -74,9 +74,9 @@ public class SQLStatementCountValidator {
      *
      * @param expectedUpdateCount expected update statement count
      */
-    public static void assertUpdateCount(int expectedUpdateCount) {
+    public static void assertUpdateCount(long expectedUpdateCount) {
         QueryCount queryCount = QueryCountHolder.getGrandTotal();
-        int recordedUpdateCount = queryCount.getUpdate();
+        long recordedUpdateCount = queryCount.getUpdate();
         if (expectedUpdateCount != recordedUpdateCount) {
             throw new SQLUpdateCountMismatchException(expectedUpdateCount, recordedUpdateCount);
         }
@@ -87,9 +87,9 @@ public class SQLStatementCountValidator {
      *
      * @param expectedDeleteCount expected delete statement count
      */
-    public static void assertDeleteCount(int expectedDeleteCount) {
+    public static void assertDeleteCount(long expectedDeleteCount) {
         QueryCount queryCount = QueryCountHolder.getGrandTotal();
-        int recordedDeleteCount = queryCount.getDelete();
+        long recordedDeleteCount = queryCount.getDelete();
         if (expectedDeleteCount != recordedDeleteCount) {
             throw new SQLDeleteCountMismatchException(expectedDeleteCount, recordedDeleteCount);
         }
